code review: Fix P1/P2 issues from code review round 5
- Fix timestamp type inconsistency in ChatMessage (i64 -> u64)
- Replace test panics with assert! in client network tests (10 locations)
- Fix unused variable warning in main.rs (_action_msg)
- Fix clippy len() >= 1 warning in game.rs (!is_empty())
- Note: Mutex poisoning not applicable (uses parking_lot::Mutex)

Test results: 74 passed, 0 failed

+++ poker_server/src/server.rs
@@ -373,7 +373,7 @@ impl PokerServer {
                         .map(|p| p.name.clone())
                         .unwrap_or_default(),
                     text,
-                    timestamp: chrono::Utc::now().timestamp(),
+                    timestamp: chrono::Utc::now().timestamp() as u64,
                 };
                 if let Err(e) = self.tx.send(ServerMessage::Chat(chat_msg)) {
                     error!("Failed to send chat message to broadcast channel: {}", e);
